Get configuration from controller.
There's an undocumented command that requests the controller
configuration data.  The configuration data holds the number of zones
configured, the number of sources configured. It includes information
about each zone; including the zone name.  It includes information
about each source; includeing the source name.

It also includes the 10 user defined custom names that can be used
for either zone names or source names.

The names are actually stored in the data as index into string
tables.  The zone and source tables have been added to the rnet
message module.

Also, the configuration data is sent in multiple messages (345 of
them). So the message decode function has been updated to handle
multi-packet messages.

Some messages from the controller should be acknowledged. If they
aren't, there will be a delay while the controller wait or, in the
case of multi-packet messages, you just get the first packet. So
add the acknowledge handshake to the TCP recieive loop.