Attempt to detect chained controllers.
Controllers can be "chained" together so that all controll is
routed through the main controller.  However, each controller
has a unique controller ID used to address commands to it.  We
could make the user enter the ID's of each connected controller,
but it would be more user friendly if we could detect this and
set it up automatically.

So we'll try querying each controller ID and see that happens.

For RIO, we can query for the controller type and we expect an
error if it doesn't exist.

For RNET, we don't know what happens if query for a non-existent
controller ID yet.