Comment out source selection triggers.
When we only had to deal with 6 sources, we had them listed
in the main control node.  When a source was deactivated or
activated, we could set a driver in the control node
appropriately.  This could then be used as a trigger.

However, with multi-controllers, we may have one controller
as the main control point, but secondary controllers with
additional sources.  We don't really know how many sources
are in-use now.

The easy thing it to just not do this anymore.