Continue working to get RNET commands/status working
Some of the get info message parsing was still wrong.  The
index in the message is different for some of the parameters.

Source indexing was still a bit off.  Should be right now.

Party Mode has 3 states (on/off/master)

When the keypad is used to change source or volume, we need
to query to see what the current source and current volume
level is.