IoX firmware bug work-a-round
The IoX does not currently support commands with parameters
per the node server API.  It simply ignores the parameters
and leaves them at the default values (value = 0, uom = -1)
which causes most comparisons to fail.

To work around this, the node server must create drivers
to hold the parameter values. So that the node server first
sets the driver and then sends the command.

The combination can then be used as trigger in IoX programs.

For example assume we want to send a DON with a parameter of 5.

We should be able to have an IoX program that simply checks

    if CTRL DON 5

But this won't work because the 5 is ignored.

So instead we need

    if CTRL DON
    AND DON is 5

So that when the DON command is sent, we then check the
DON driver status for 5.