Add commands to controller node for all zones on/off
The all zones on/off is triggered by sending an event.  For RNET
we create a new command that formats the data for an event and
sends it.  For RIO, we use the set_param() call and add a
parameter for this using the same format we use for things like
party mode on/off.  However, since this is for all zones and not
a single zone, what should the zone be set to? Unknown if this
will work.